cpufreq/amd-pstate: Convert all perf values to u8
Mainline: 555bbe67a622b297405e246d1868dbda3284bde8
From: v6.15-rc1
Severity: Moderate

task: 564669

All perf values are always within 0-255 range, hence convert their
datatype to u8 everywhere.

Link: https://lore.kernel.org/r/20250205112523.201101-7-dhananjay.ugwekar@amd.com

(backported from commit 555bbe67a622b297405e246d1868dbda3284bde8)

K2CI-Arch: Amd64
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/153131
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>

+++ drivers/cpufreq/amd-pstate.h
@@ -69,13 +69,13 @@ struct amd_cpudata {
 	struct	freq_qos_request req[2];
 	u64	cppc_req_cached;
 
-	u32	highest_perf;
-	u32	nominal_perf;
-	u32	lowest_nonlinear_perf;
-	u32	lowest_perf;
-	u32     prefcore_ranking;
-	u32     min_limit_perf;
-	u32     max_limit_perf;
+	u8	highest_perf;
+	u8	nominal_perf;
+	u8	lowest_nonlinear_perf;
+	u8	lowest_perf;
+	u8	prefcore_ranking;
+	u8	min_limit_perf;
+	u8	max_limit_perf;
 	u32     min_limit_freq;
 	u32     max_limit_freq;
 
@@ -92,10 +92,10 @@ struct amd_cpudata {
 	bool	hw_prefcore;
 
 	/* EPP feature related attributes*/
-	s16	epp_cached;
+	u8	epp_cached;
 	u32	policy;
 	bool	suspended;
-	s16	epp_default;
+	u8	epp_default;
 };
 
 /*
